* Moved tha AccessLevel class + Added support for a server obfuscation table
