+++ res/conf/minecraft_server.obf
@@ -1,0 +1,1 @@
+# It's lonely in here :,c
