+++ src/com/blazeloader/bl/obf/AccessLevel.java
@@ -1,4 +1,4 @@
-package com.blazeloader.util.transformers;
+package com.blazeloader.bl.obf;
 
 import static org.objectweb.asm.Opcodes.*;
